ChampionInfo.js의 챔피언리스트를 외부파일로 분리
