+++ src/pages/ChampionInfo.js
@@ -1,6 +1,5 @@
-import axios from "axios";
-import React, { useEffect, useState } from "react";
-import { useNavigate } from "react-router-dom";
+import React from "react";
+import ChampionListComponent from "../components/contents/ChampionInfo/ChampionListComponent";
 import RotationChampions from "../components/contents/ChampionInfo/RotationChampions";
 import styles from "./ChampionInfo.module.css";
 
@@ -9,10 +8,8 @@ const ChampionInfo = (
     championData,
     onErrorImg,
 ) => {
-    const values = Object.values(championData.championData[0])
-    const championObjectArray = Object.values(values[3]) // 챔피언정보가 담긴 객체들의 배열
-
-    const navigate = useNavigate();
+    const values = Object.values(championData.championData[0]); // 데이터 가공을 위해 value값만 뽑아냄
+    const championObjectArray = Object.values(values[3]); // 챔피언정보가 담긴 객체들의 배열
 
     return (
         <>
@@ -26,23 +23,10 @@ const ChampionInfo = (
             </div>
             {/* 챔피언 리스트 */}
             <div className={styles['championInfo-container']}>
-                {championObjectArray.map((championName, index) => {
-                    return (
-                        <div 
-                            className={styles['championInfo-champContainer']} 
-                            key={index}
-                            onClick={() => {navigate(`/championInfo/${championObjectArray[index].key}`)}}
-                        >
-                            <img 
-                                src={`https://ddragon.leagueoflegends.com/cdn/12.21.1/img/champion/${championObjectArray[index].id}.png`} 
-                                className={styles['championInfo-img']}
-                                onError={() => onErrorImg()}
-                                alt={championObjectArray[index].id}
-                            />
-                            <div>{championObjectArray[index].name}</div>
-                        </div>
-                    )})
-                }
+                <ChampionListComponent 
+                    onErrorImg={onErrorImg} 
+                    championObjectArray={championObjectArray}
+                />
             </div>
         </>
     )
